webkit/data/layout_tests/chrome/fast/dom/messageport-gc-expected.txt

ALERT: port1
ALERT: from port2
ALERT: port2
ALERT: from port1
